# docs/swagger.yaml
definitions:
...
  dto.GenreDTO:
    properties:
      id:
        type: integer
      name:
        type: string
      rusName:
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/genres/album/{albumID}:
    get:
      description: Retrieves a list of all genres for a given album ID from the database.
      parameters:
      - description: Album ID
        in: path
        name: albumID
        required: true
        type: integer
      responses:
        "200":
          description: List of genres by album
          schema:
            items:
              $ref: '#/definitions/dto.GenreDTO'
            type: array
        "404":
          description: No genres found for the album
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to load genres
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get all genres by album ID
  /api/v1/genres/all:
    get:
      description: Retrieves a list of all genres from the database.
      responses:
        "200":
          description: List of all genres
          schema:
            items:
              $ref: '#/definitions/dto.GenreDTO'
            type: array
        "404":
          description: No genres found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to load genres
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get all genres
  /api/v1/genres/artist/{artistID}:
    get:
      description: Retrieves a list of all genres for a given artist ID from the database.
      parameters:
      - description: Artist ID
        in: path
        name: artistID
        required: true
        type: integer
      responses:
        "200":
          description: List of genres by artist
          schema:
            items:
              $ref: '#/definitions/dto.GenreDTO'
            type: array
        "404":
          description: No genres found for the artist
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to load genres
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get all genres by artist ID
  /api/v1/genres/track/{trackID}:
    get:
      description: Retrieves a list of all genres for a given track ID from the database.
      parameters:
      - description: Track ID
        in: path
        name: trackID
        required: true
        type: integer
      responses:
        "200":
          description: List of genres by track
          schema:
            items:
              $ref: '#/definitions/dto.GenreDTO'
            type: array
        "404":
          description: No genres found for the track
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to load genres
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get all genres by track ID
...
swagger: "2.0"